Extract game screen to common
