+++ front/src/common/game_screen.scss
@@ -1,0 +1,18 @@
+.game-board {
+    .lifes-display {
+      position: absolute;
+      top: 15px;
+      left: 20px;
+    }
+  
+    .remaining-questions {
+      position: absolute;
+      top: 15px;
+      right: 15px;
+    }
+  
+    .question {
+      margin-top: 50px;
+    }
+  }
+
